% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofsaa.R
\name{formatColumns}
\alias{formatColumns}
\title{Format values according to type representation for OFSAA}
\usage{
formatColumns(value, data_type = c("varchar", "number", "date", "?"),
  max_num_len = NULL, max_dec_len = NULL)
}
\arguments{
\item{value}{vector of values}

\item{data_type}{data type of values in a vector}

\item{max_num_len}{maximum length of number}

\item{max_dec_len}{maximum number of decimal digits in number}
}
\description{
Function takes vector of values as their type and returns vector of formatted
values ready to be stored in a file
}
