% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofsaa.R
\name{ofsaa_data}
\alias{ofsaa_data}
\title{Create OFSAA data object}
\usage{
ofsaa_data(dataframe, columns = NULL)
}
\arguments{
\item{dataframe}{dataframe to be arranged}

\item{columns}{vector of column names that defines what columns should
present in the result dataframe and in which order should they appear}
}
\description{
Expand and arrange columns into specific order. Function takes dataframe and desired list of columns and returns dataframe
with this columns arranged accordingly if column present and original
dataframe and missing in columns than it will be dropped. If column is
missing in original dataframe and present in columns than new column with NA
values will be generated.
}
