% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofsaa.R
\name{ofsaa_data_object}
\alias{ofsaa_data_object}
\title{generate vector of lines to be stores in a file}
\usage{
ofsaa_data_object(header, data, footer)
}
\arguments{
\item{header}{header object created by \link{ofsaa_header} function}

\item{data}{data object created by \link{ofsaa_data} function}

\item{footer}{footer object created by \link{ofsaa_footer} function}
}
\description{
generate vector of lines to be stores in a file
}
