% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofsaa.R
\name{ofsaa_footer}
\alias{ofsaa_footer}
\title{Create OFSAA footer object}
\usage{
ofsaa_footer(nrow)
}
\arguments{
\item{nrow}{number of rows to be reported in footer (actually number of rows
in body dataframe)}
}
\description{
Function creates footer object which then can be serialized (written as file)
}
