% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofsaa.R
\name{ofsaa_header}
\alias{ofsaa_header}
\title{Create OFSAA header object}
\usage{
ofsaa_header(fileset, csa_id, csa_name, version, business_date,
  business_time)
}
\arguments{
\item{fileset}{File Set code}

\item{csa_id}{CSA Table ID (like 1018, 1006, etc)}

\item{csa_name}{SCA Table Name (like STG_LOAN_CONTRACTS)}

\item{version}{version}

\item{business_date}{business date in YYYYMMDD text format}

\item{business_time}{in HH24miss format}
}
\description{
Function creates header object which then can be serialized (written as file)
}
