% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofsaa-writer.R
\name{write_ofsaa}
\alias{write_ofsaa}
\title{write OFSAA data object into file}
\usage{
write_ofsaa(data_object, file)
}
\arguments{
\item{data_object}{object produced by \link{ofsaa_data_object} function}

\item{file}{filename to store data}
}
\description{
write OFSAA data object into file
}
